[dts pinenote.dtsi] switch touchscreen (cyttsp5) interrupt from LEVEL LOW to EDGE_FALLING.
My notes:

- I think the interrupt should indeed be FALLING EDGE, and the level low
  interrupt type documented here:

https://github.com/torvalds/linux/blob/63355b9884b3d1677de6bd1517cd2b8a9bf53978/Documentation/devicetree/bindings/input/touchscreen/cypress%2Ctt21000.yaml#L86

is a remnant from the android cyttsp5 driver, which in turn is a remnant
of the android cyttsp4 driver.

The android driver has this if-statement where it basically supports
both level and falling edge interrupts:

https://github.com/aosp-rockchip/android_kernel_rockchip_rk356x/blob/8da04464e2bf51e45673d264c7a16e7309b8c793/drivers/input/touchscreen/cyttsp5/cyttsp5_core.c#L5962

However, when I grep for this level_irq_udelay variable, it is not used
anywhere in the cyttsp5 android driver. However, the cyttsp4 driver
actually uses it to add a delay to the interrupt handler, which makes
sense when you trigger on level low:

https://github.com/aosp-rockchip/android_kernel_rockchip_rk356x/blob/8da04464e2bf51e45673d264c7a16e7309b8c793/drivers/input/touchscreen/cyttsp4_core.c#L1238

Summarizing, I think the dt scheme and the pinenote dts should be fixed.
In case that you agree with my assessment, could you by chance advise me
how to proceed?