drm: fix loading of epd waveforms
Fix a bug in the drm epd helper library that basically imported waveform
data in a transposed form. This lead to nearly-inverted waveforms (exact
for A2, but not the others) and a bug in the rockchip_ebc driver that
had the next/prev hardware buffers switched.

Unfortunately, importing the waveform got slightly more complicated, as
importing to 5BIT-Packed format is not as simple as copying the
corresponding bytes anymore. Therefore, the default decoding format was
switched to 5BIT (unpacked).